requirements.txt

Flask
uWSGI
flask_restful
flask_sqlalchemy
PyJwt
sqlalchemy==1.4.44
geoalchemy2==0.10.2
psycopg2-binary
sqlalchemy_utils
pytest
pytest-celery
pydantic
flask_cors

msgpack-python
celery[sqs,sqlalchemy]


matplotlib
fiona
pygeos
Shapely
openrouteservice
pandas==1.5.1
geopandas
numpy
pyproj

folium
json2html